requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
prometheus-client==0.20.0
opentelemetry-sdk==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-requests==0.47b0
opentelemetry-exporter-otlp==1.26.0
requests==2.32.3
